requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
icrawler==0.6.7
idna==3.4
lxml==4.9.3
Pillow==10.0.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
urllib3==2.0.4
